Yet another respectable performance improvement (~10%) by tuning the thread count
It turns out using the number of logical cores is the fastest. I'm guessing tokio's high default thread count (512) is to accommodate things like network requests where threads can go to sleep for a while. In our case, I/O is presumably fast enough that the threads are constantly thrashing.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>