Rewrite scheduling algorithm in preparation for supporting exact num files
This implementation uses depth-first traversal which is bad. Ideally, we'd use breadth-first, but doing so scales memory with the width of the tree which in our case can be very large. This implementation is currently a few percent to a few dozen percent slower than the previous version which sucks — the next step is to try combining depth-first and breadth-first into some sort of batched breadth-first algorithm with bounded memory.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>