Massively reduce memory usage by continuously draining tasks
With 100K files, memory usage drops from 1,815,704 bytes to 627,966 bytes.

The task capacity was completely wrong since it used the number of files (should have been num_files / files_per_dir). It's now some multiple of the number of CPU cores since that's what limits how many tasks can run in parallel. The idea is to drain as many tasks as possible to minimize the memcpy moving the remaining tasks to the front, but not so much that the machine ever ends up with spare capacity.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>